Move link logic into a separate function
The (vim)-like link logic is moved into an own separate function.
That had to be done because if MyBrowse was called like

    $ mybrowse /usr/bin

it did not show a file index instead because link logic was only called
if an URL has been typed into addressbar.

Three commands are not migrated to the function:
 * :back, :undo, :u
 * :forward, :redo, :r
 * :reload, :rl

because it does not make sense to call these from the command line.